// control-plane/mesh-ui/src/api/openapi/types.gen.ts
export type V1Node = {
  id?: string
  machineKey?: string
  nodeKey?: string
  discoKey?: string
  ipAddresses?: Array<string>
  name?: string
  user?: V1User
  lastSeen?: string
  expiry?: string
  preAuthKey?: V1PreAuthKey
  createdAt?: string
  registerMethod?: V1RegisterMethod
  /**
   * Deprecated
   * repeated string forced_tags = 18;
   * repeated string invalid_tags = 19;
   * repeated string valid_tags = 20;
   */
  givenName?: string
  online?: boolean
  approvedRoutes?: Array<string>
  availableRoutes?: Array<string>
  subnetRoutes?: Array<string>
  tags?: Array<string>
}

export type V1PreAuthKey = {
  user?: V1User
  id?: string
  key?: string
  reusable?: boolean
  ephemeral?: boolean
  used?: boolean
  expiration?: string
  createdAt?: string
  aclTags?: Array<string>
}

export type V1RegisterMethod =
  | 'REGISTER_METHOD_UNSPECIFIED'
  | 'REGISTER_METHOD_AUTH_KEY'
  | 'REGISTER_METHOD_CLI'
  | 'REGISTER_METHOD_OIDC'
// ...
export type V1User = {
  id?: string
  name?: string
  createdAt?: string
  displayName?: string
  email?: string
  providerId?: string
  provider?: string
  profilePicUrl?: string
}